Initialize worker if plugin provides necessary interface
This change ignores the schedule feature and initializes the worker(s) if the
plugin provides the worker interface even if it doesn't list the schedule
feature as supported.  This relaxes support for sloppy plugins that don't list
their features, and technically allows a plugin to only provide the interface
and use it for synchronous thread-safe restore only (not that this makes much
sense).