Use sigaction() instead of signal()
Issue is that even after we do ctrl+c, fgets is waiting for a newline.

Ref : man of signal(7)
If a signal handler is invoked while a system call or library function call is blocked, then either:
1. the call is automatically restarted after the signal handler returns
2. the call fails with the error EINTR.
Which of these two behaviors occurs depends on the interface and whether or not the signal handler was established using the SA_RESTART flag (see sigaction(2)).

We have to use either the siginterrupt() function together with signal(), or use sigaction() instead of signal() for registering the signal handler,
in order to disable restarting a read() system call after a signal.

Signed-off-by: Laxmi Devi <Laxmi.Devi@in.bosch.com>
Signed-off-by: Timo Wischer <twischer@de.adit-jv.com>