worker: Do not read response when not yet available
Without this patch the size of the response might be read successfully
but the response will not be read. Therefore the while loop would exit
and when entering the next time the size will be read again.
But this second read of size will actually already read data from the
response. Therefore the response will be corrupted and the response
buffer cannot sync again.

To avoid this issue it will first be checked if there is enough data
available in the ring buffer to read the size and response. If not try
again later.

Signed-off-by: Timo Wischer <twischer@de.adit-jv.com>