Remove transport position dumping from Jack process callback
This is handy at times, but prints in the process callback which is obviously
not realtime safe.  The communication dumping is, though, and both are
controlled by the same option, so there wasn't a way to enable that without
introducing a pretty severe performance issue in the audio thread.

So, simply remove it for now.  If transport dumping is really needed, it can be
implemented by sending the information to the main thread for printing there,
which will be easier after the upcoming work to improve the communication
facilities.