Clean up include paths
Removes the source directory from the include path flags passed to the
compiler, so that quoted includes always refer to files relative to the one
they're in, and angled includes never refer to anything in the source tree.
This fixes potential clashes between the name of headers here and on the
system.