Rename jalv_internal.h to jalv.h
The "internal" name never made much sense to begin with (since this is a
program, not a library), but now this header only describes the "main" Jalv
struct and functions directly associated with it (the top-level interface of
the application), so name it accordingly.

This also makes include-what-you-use actually check this header, which it
wasn't before for some mysterious reason.