jalv, state: Set default values when control is unset in state
When loading a preset (Lilv speak: state), controls whose value is
not set in the preset are set to their default values. This is to
get a consistent state of the controls even if certain parameters
are missing in the preset file. This can happen for instance if the
plugin is updated and new controls added, but the preset file to be
loaded was createed by the older version of the plugin.

In this commit, this is done unconditionally. This may not be
optimal for plugins where it is desired to load 'partial presets'
which don't contain values for all parameters.