jalv, jalv_console, state: Make default setting configurable
Make setting of parameters to their default values when missing
from loaded preset files optional. Either set the -D flag on the
command line when starting, which affects the behavior of the
'preset' CLI command, or use the 'presetd' command to force
the feature regardless of the presence of the -D flag.
In contrast, the 'presetn' command always disregards the
-D flag it and never sets unused parameters to their defaults.

The -D flag also affects how the graphic versions of jalv such as
jalv.gtk3 behave when loading preset files.